requirements.txt
dspy-ai>=2.4.0
datasets>=2.14.0
huggingface_hub>=0.19.0
pandas>=2.0.0
python-dotenv>=1.0.0

